refine build scripts

+++ build/prebuild-download.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 
-# Copyright (c) 2021- KanTV Authors
-
 # Description: download Android SDK and Android NDK and HEXAGON_TOOLs_8.8.06.tar.gz for build the entire project in command-line mode
+
+# verified on Ubuntu 20.04, Ubuntu 24.04
 
 set -e
 
@@ -32,14 +32,16 @@ if [ ! -f ${ANDROID_NDK}/build/cmake/android.toolchain.cmake ]; then
     is_android_ndk_exist=0
 fi
 
-if [ ! -f ${PROJECT_ROOT_PATH}/prebuilts/toolchain/android-sdk/cmdline-tools/bin/sdkmanager ]; then
+if [ ! -f ${PROJECT_ROOT_PATH}/prebuilts/toolchain/android-sdk/cmdline-tools/latest/bin/sdkmanager ]; then
     echo -e "${TEXT_RED}Android SDK cmdline-tools not exist, pls check...${TEXT_RESET}\n"
     is_cmdlinetools_exist=0
 fi
 
 if [ ! -f ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools/8.8.06/NOTICE.txt ]; then
-    echo -e "${TEXT_RED}HEXAGON_TOOLS not exist, pls check...${TEXT_RESET}\n"
+    echo -e "${TEXT_RED}hexagon LLVM toolchain not exist, pls check...${TEXT_RESET}\n"
     is_hexagon_llvm_exist=0
+else
+    echo -e "${TEXT_RED}hexagon LLVM toolchain already exist${TEXT_RESET}\n"
 fi
 
 if [ ${is_android_ndk_exist} -eq 0 ]; then
@@ -93,32 +95,37 @@ else
 fi
 
 
-#prepare for cmdline build
+if [ ${is_cmdlinetools_exist} -eq 0 ]; then
+    #prepare for cmdline build
+    export ANDROID_HOME=${PROJECT_ROOT_PATH}/prebuilts/toolchain/android-sdk/
+    export PATH=${ANDROID_HOME}/cmdline-tools/latest/bin:${PATH}
 
-export ANDROID_HOME=${PROJECT_ROOT_PATH}/prebuilts/toolchain/android-sdk/
-export PATH=${ANDROID_HOME}/cmdline-tools/latest/bin:${PATH}
-
-# check sdkmanager works
-sdkmanager --version
-if [ $? -ne 0 ]; then
-    printf "android cmdline-tools could not work properly, pls check development envs\n"
-    exit 1
+    # check sdkmanager works
+    sdkmanager --version
+    if [ $? -ne 0 ]; then
+        printf "android cmdline-tools could not work properly, pls check development envs\n"
+        exit 1
+    fi
+    yes | sdkmanager --licenses
+    yes | sdkmanager --install "platforms;android-34"
+    yes | sdkmanager --install "build-tools;34.0.0"
+    yes | sdkmanager --install "cmake;3.22.1"
+else
+    printf "android sdk already exist \n\n"
 fi
-
-yes | sdkmanager --licenses
-yes | sdkmanager --install "platforms;android-34"
-yes | sdkmanager --install "build-tools;34.0.0"
-yes | sdkmanager --install "cmake;3.22.1"
 
 
 #download customized LLVM toolchain HEXAGON_TOOLs_8.8.06.tar.gz
 if [ ${is_hexagon_llvm_exist} -eq 0 ]; then
-    echo -e "begin downloading HEXAGON_TOOLs(a customized LLVM toolchain) \n"
+    echo -e "begin downloading hexagon LLVM toolchain \n"
     wget --no-config --quiet --show-progress -O ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools/HEXAGON_TOOLs_8.8.06.tar.gz https://github.com/kantv-ai/toolchain/raw/refs/heads/main/HEXAGON_TOOLs_8.8.06.tar.gz
     if [ $? -ne 0 ]; then
-        printf "failed to download HEXAGON_TOOLs(a customized LLVM toolchain)\n"
+        printf "failed to download hexagon LLVM toolchain\n"
         exit 1
     fi
 
-    zcat ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools/HEXAGON_TOOLs_8.8.06.tar.gz | tar -C ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools -xf -
+    zcat ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools/HEXAGON_TOOLs_8.8.06.tar.gz | tar -C ${PROJECT_ROOT_PATH}/prebuilts/Hexagon_SDK/6.2.0.1/tools/HEXAGON_Tools -xvf -
+    printf "install hexagon LLVM toolchain successfully\n\n"
+else
+    printf "hexagon LLVM toolchain already exist\n\n"
 fi
